% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\docType{package}
\name{browse-package}
\alias{browse}
\alias{browse-package}
\title{browse: A file browser for RStudio}
\description{
Implements a file browser within RStudio using `finder.js` 
    which supports multi-column view. A key feature of the package is 
    that it supports re-mapping of the files without changing the 
    directory structure.
}
\keyword{internal}
