% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\alias{link}
\title{Navigate and Link to Remote Git Files}
\usage{
browse(path = NULL, remote = "origin")

link(path = NULL, remote = "origin")
}
\arguments{
\item{path}{path to the file. The path may be relative to the current
working directory, or an absolute path.
The default, \code{path = NULL} works on a RStudio editor selection.}

\item{remote}{Name of the remote to link to.
Optionally, you can change by setting
\code{options(browse.remote.default = "remote_name")}.}
}
\description{
\code{link} creates a permalink to the remote file and copies it to your
clipboard for easy sharing. This permalink is based on the HEAD commit
of the local git repo.

\code{browse} does the same thing, but also opens the website in your default
web browser.

Works with Github, Gitlab, and Bitbucket repos.
}
\details{
In the \code{path} argument, you can optionally include line numbers after
the path to the file. See the examples section below. Formatting varies
between Github, Gitlab, and Bitbucket.
}
\examples{
\dontrun{
# works on an RStudio selection in the editor
browse()
link()

# working directory is the top level of repo
browse("R/browse.R")
link("R/browse.R#L6-L9")

# working directory is the R/ folder of repo
link("../README.md")

# relative or absolute paths to other repos
browse("~/rpackages/dplyr/DESCRIPTION")
}

}
