% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_app.R
\name{build_app}
\alias{build_app}
\title{Create an app from a html object of class "shiny.tag"}
\usage{
build_app(html, user_function = identity)
}
\arguments{
\item{html}{A html object of class "shiny.tag".}

\item{user_function}{A R function that takes and returns a named list.}
}
\description{
Create an app from a html object of class "shiny.tag"
}
