% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_directory.R
\name{dir_to_JSON}
\alias{dir_to_JSON}
\title{Directory to JSON}
\usage{
dir_to_JSON(path, skip_first_level = F)
}
\arguments{
\item{path}{character; the path to the directory}

\item{skip_first_level}{T or F; should the first level be skipped?}
}
\description{
Directory to JSON
}
