% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_directory.R
\name{directory}
\alias{directory}
\title{Browse directory}
\usage{
directory(path, reminder = T)
}
\arguments{
\item{path}{character; the path to the directory}

\item{reminder}{T or F; should USAGE be displayed?}
}
\description{
Browse directory
}
