% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{github_url}
\alias{github_url}
\title{Format Github}
\usage{
github_url(path)
}
\arguments{
\item{path}{The relative path of the file from the top of the git repo.}
}
\description{
Formats Github URL based on the repo in the current working directory.
Uses the current HEAD of the git repo.
}
\examples{
github_url("R/browse.R#L6-L9")
github_url("R/browse.R#L6")

# opens URL in default browser
\dontrun{
github_url("R/install.R") \%>\%
  browseURL()
}

}
