% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_directory.R
\name{JSON_to_app}
\alias{JSON_to_app}
\title{Launch app from JSON}
\usage{
JSON_to_app(json, reminder = T)
}
\arguments{
\item{json}{character; the JSON string containing the data.}

\item{reminder}{T or F; should USAGE be displayed?}
}
\description{
Launch app from JSON
}
