% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{link}
\alias{link}
\title{Link to Remote File}
\usage{
link(path = NULL, remote = "origin")
}
\arguments{
\item{path}{The relative path of the file from the top of the git repo.
The default, \code{path = NULL} works on a RStudio selection.}

\item{remote}{Name of the remote to link to.
Optionally, you can change by setting
\code{options(browse.remote.default = "remote_name")}.}
}
\description{
Creates a Github permalink based on the repo in the current working directory.
Uses the current HEAD of the git repo.
}
\examples{
\dontrun{
link("R/browse.R#L6-L9")
link("R/browse.R#L6")

# works on a RStudio selection
link()
}

}
