% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_dir_class.R
\docType{data}
\name{virtual_dir}
\alias{virtual_dir}
\title{R6 virtual directory object class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
virtual_dir
}
\description{
R6 virtual directory object class
}
\keyword{datasets}
