% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\docType{package}
\name{browse-package}
\alias{browse}
\alias{browse-package}
\title{browse: A file browser for RStudio}
\description{
Implements a file browser using `finder.js` to be used 
    temporarily until RStudio changes to (hopefully) a multi-column 
    file browser. An important feature of the package is that it 
    supports re-mapping of the files without changing the directory 
    structure.
}
\keyword{internal}
