% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{link}
\alias{link}
\title{Link to Remote File}
\usage{
link(path = NULL)
}
\arguments{
\item{path}{The relative path of the file from the top of the git repo.
The default, \code{path = NULL} works on a RStudio selection.}
}
\description{
Creates a Github permalink based on the repo in the current working directory.
Uses the current HEAD of the git repo.
}
\examples{
link("R/browse.R#L6-L9")
link("R/browse.R#L6")

# works on RStudio selection
link()

# opens URL in default browser
\dontrun{
link("R/install.R") \%>\%
  browseURL()
}

}
